chore: move utilities to `FormUtil.js`
